requirements.txt
rembg==2.0.57
diffusers==0.29.2 
trimesh==4.4.9 
pyrender==0.1.45 
Pillow==10.4.0 
numpy==1.26.4 
torch==2.4.1